added buttons for navigation
