+++ src/components/chatComponents/Navbar.jsx
@@ -12,7 +12,7 @@ const Navbar = () => {
     <div className='navbar'>
       <span className="logo">Pixel Palz</span>
       <div className="user">
-        <img src='https://t4.ftcdn.net/jpg/05/49/98/39/360_F_549983970_bRCkYfk0P6PP5fKbMhZMIb07mCJ6esXL.jpg' alt='profile picture' onClick={() => navigate('shop_page')}/>
+        <img src='https://t4.ftcdn.net/jpg/05/49/98/39/360_F_549983970_bRCkYfk0P6PP5fKbMhZMIb07mCJ6esXL.jpg' alt='profile picture' onClick={() => navigate('profile')}/>
         <span>{currentUser.displayName}</span>
         <button onClick={()=>signOut(auth)}>Logout</button>
       </div>
