+++ src/style.scss
@@ -343,11 +343,12 @@ body {
 }
 
 .profile{
-    background-color: #F5EBE0;
+    background-color: #faebd7;
     height: 100vh;
     display: flex;
     align-items: center;
     justify-content: center;
+    font-family: 'pixelFont';
 
     h1{
         color:black;
@@ -360,7 +361,8 @@ body {
     }
 
     .container{
-        background-color: #D9D9D9;
+        background-color: #faebd7;
+        border: 1px solid #a18585;
         height:80%;
         width:35%;
         border-radius: 30px;
@@ -369,10 +371,48 @@ body {
         align-items: center;
         display: flex;
         flex-direction: column;
-        
+
+        .bar{
+            display: flex;
+            align-items: center;
+        }
+
+        .button2:hover,
+        .button:hover {
+        background-color: #c6aeae;
+        cursor: pointer;
+        }
+
+        .button {
+            background-color: #a18585;
+            padding:10px;
+            border-radius:30px;
+            position:absolute;
+            left:20px;
+    
+            a {
+                text-decoration: none;
+                color: black;
+            }
+        }
+
+        .button2 {
+            background-color: #a18585;
+            padding:10px;
+            border-radius:30px;
+            position:absolute;
+            right:20px;
+    
+            img{
+                height: 25px;
+                width: 30px;
+                object-fit: cover;
+                cursor: pointer;
+            }
+        }
         
         .circle{
-            background-color: #504a4a;
+            background-color: #a18585;
             height:300px;
             width:300px;
             border-radius:50px;
@@ -388,7 +428,8 @@ body {
         }
     }
     .container2{
-        background-color: #D9D9D9;
+        background-color: #faebd7;
+        border: 1px solid #a18585;
         height:80%;
         width:45%;
         border-radius: 30px;
@@ -397,6 +438,20 @@ body {
         align-items: center;
         display: flex;
         flex-direction: column;
+
+        .button:hover {
+            background-color: #c6aeae;
+            cursor: pointer;
+        }
+
+        .button {
+            background-color: #a18585;
+            padding:10px;
+            border-radius:30px;
+            position:absolute;
+            top:25px;
+            left:20px;
+        }
 
         table{
             border-collapse: collapse;
@@ -408,14 +463,14 @@ body {
         }
 
         th{
-            background-color: grey;
+            background-color: #a18585;
             text-align:center;
             font-weight:1;
             padding:8px;
         }
 
         td{
-            background-color: white;
+            background-color: #e1d3c1;
             height:100px;
             width:100px;
             button{
